Imporve Notifcations, add missing descriptions
At date of this commit Notify with it's last commit got incompatible when ustilizing multigrid
in nvim GUIs (tested with neovide and nvui). A temporary fix is to
revert to the last commit in the notify plugin.
use { "rcarriga/nvim-notify", commit = "6f298f219f373b657b40dab7468bbd19e9ba8159" }